Broke Application.cs up and put into subfolder
